move RenderCallback,RenderDbGelper, RenderService to model package; add dependency RecycleView
