+++ app/src/main/java/com/tt/reminder/fragment/EditTaskFragment.java
@@ -17,7 +17,7 @@ import com.tt.sharedbaseclass.constant.Constant;
 import com.tt.sharedbaseclass.fragment.EditTaskFragmentBase;
 import com.tt.sharedbaseclass.model.RenderObjectBeans;
 import com.tt.sharedbaseclass.model.TaskBean;
-import com.tt.sharedbaseclass.service.RenderCallback;
+import com.tt.sharedbaseclass.model.RenderCallback;
 import com.tt.sharedbaseclass.view.WheelView;
 
 import java.util.Calendar;
